requirements.txt
# how to instal? 
# version of packages: %pip install -r requirements.txt
absl-py==1.0.0
accelerate==0.21.0
adagio==0.2.6
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
autogluon==1.1.1
autogluon.common==1.1.1
autogluon.core==1.1.1
autogluon.features==1.1.1
autogluon.multimodal==1.1.1
autogluon.tabular==1.1.1
autogluon.timeseries==1.1.1
autots==0.6.15
azure-core==1.29.1
azure-cosmos==4.3.1
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==4.1.0
blinker==1.4
blis==0.7.11
boto3==1.24.28
botocore==1.27.96
cachetools==5.3.2
catalogue==2.0.10
catboost==1.2.7
category-encoders==2.6.3
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpathlib==0.16.0
cloudpickle==2.0.0
cmdstanpy==1.2.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm==0.1.2
confection==0.1.4
configparser==5.2.0
contourpy==1.0.5
coreforecast==0.0.15
crcmod==1.7
cryptography==39.0.1
cycler==0.11.0
cymem==2.0.8
Cython==0.29.32
dacite==1.8.1
databricks-automl-runtime==0.2.20
databricks-cli==0.18.0
databricks-feature-engineering==0.2.1
databricks-sdk==0.1.6
dataclasses-json==0.6.3
datasets==2.15.0
dbl-tempo==0.1.26
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.12.4
defusedxml==0.7.1
dill==0.3.6
diskcache==5.6.3
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
docstring-to-markdown==0.11
entrypoints==0.4
evaluate==0.4.1
executing==0.8.3
facets-overview==1.1.1
fastai==2.7.18
fastcore==1.7.20
fastdownload==0.0.7
fastjsonschema==2.19.1
fastprogress==1.0.3
fasttext==0.9.2
filelock==3.14.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.25.0
frozenlist==1.4.1
fs==2.4.16
fsspec==2023.6.0
fugue==0.9.1
future==0.18.3
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.27
gluonts==0.15.1
google-api-core==2.15.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.3
greenlet==2.0.1
grpcio==1.48.2
grpcio-status==1.48.1
gunicorn==20.1.0
gviz-api==1.10.0
h5py==3.7.0
hierarchicalforecast==0.4.2
hjson==3.1.0
holidays==0.38
horovod==0.28.1
htmlmin==0.1.12
httplib2==0.20.2
huggingface-hub==0.19.4
humanfriendly==10.0
hyperopt==0.2.7
idna==3.4
ImageHash==4.3.1
imageio==2.36.0
imbalanced-learn==0.11.0
importlib-metadata==4.11.3
importlib-resources==6.1.1
ipykernel==6.25.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
joblibspark==0.5.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2024.10.1
jupyter-client==7.3.4
jupyter-server==1.23.4
jupyter_core==5.2.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.4
langchain==0.0.348
langchain-core==0.0.13
langcodes==3.3.0
langsmith==0.0.79
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==15.0.6.1
librosa==0.10.1
lightgbm==4.1.0
lightning==2.3.3
lightning-utilities==0.11.9
linkify-it-py==2.0.3
llvmlite==0.39.1
lxml==4.9.1
Mako==1.2.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
mistune==3.0.2
ml-dtypes==0.2.0
mlflow-skinny==2.9.2
mlforecast==0.10.0
model-index==0.1.11
more-itertools==8.10.0
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.4
multimethod==1.10
multiprocess==0.70.14
murmurhash==1.0.10
mypy-extensions==0.4.3
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==7.16.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.4.2
ninja==1.11.1.1
nlpaug==1.1.11
nltk==3.7
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
nptyping==2.4.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.20.1
openai==0.28.1
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt-einsum==3.3.0
optimum==1.18.1
ordered-set==4.1.0
orjson==3.10.11
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.3
pathspec==0.10.3
patsy==0.5.3
pdf2image==1.17.0
petastorm==0.12.1
pexpect==4.8.0
phik==0.12.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
pmdarima==2.0.4
pooch==1.4.0
preshed==3.0.9
prometheus-client==0.14.1
prompt-toolkit==3.0.36
prophet==1.1.5
protobuf==4.24.0
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.4.0
py4j==0.10.9.7
pyarrow==8.0.0
pyarrow-hotfix==0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycparser==2.21
pycryptodome==3.21.0
pydantic==1.10.6
pyflakes==3.1.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pynvml==11.5.0
pyodbc==4.0.32
pyparsing==3.0.9
pyright==1.1.294
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-editor==1.0.4
python-lsp-jsonrpc==1.1.1
python-lsp-server==1.8.0
pytoolconfig==1.2.5
pytorch-lightning==2.3.3
pytorch-metric-learning==2.3.0
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
quadprog==0.1.13
ray==2.31.0
referencing==0.35.1
regex==2022.7.9
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.4.2
rope==1.7.0
rpds-py==0.21.0
rsa==4.9
s3transfer==0.6.2
safetensors==0.4.1
scikit-base==0.11.0
scikit-image==0.20.0
scikit-learn==1.5.1
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqeval==1.2.2
shap==0.44.0
simplejson==3.17.6
six==1.16.0
sktime==0.34.0
slicer==0.0.7
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spark-tensorflow-distributor==1.0.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.2.0
stanio==0.3.0
statsforecast==1.7.8
statsmodels==0.13.5
sympy==1.11.1
tabulate==0.8.10
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.14.0
tensorboardX==2.6.2.2
tensorflow-cpu==2.14.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
textual==0.78.0
thinc==8.2.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.5.2
timm==0.9.16
tinycss2==1.2.1
tokenize-rt==4.2.1
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torchmetrics==1.2.1
torchvision==0.18.1
tornado==6.1
tqdm==4.65.2
traitlets==5.7.1
transformers==4.39.3
triad==0.9.8
triton==2.3.1
typeguard==2.13.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
ujson==5.4.0
unattended-upgrades==0.1
urllib3==1.26.14
utilsforecast==0.1.9
virtualenv==20.16.7
visions==0.7.5
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
whatthepatch==1.0.2
widgetsnbextension==3.6.1
window_ops==0.0.15
wordcloud==1.9.3
wrapt==1.14.1
xgboost==1.7.6
xxhash==3.4.1
yapf==0.33.0
yarg==0.1.9
yarl==1.9.4
ydata-profiling==4.2.0
zipp==3.11.0